refactor: update imports from program-sections to sections across multiple files
